Revamping the way opcodes work
- Rather than attempting to (rather pointlessly) serialize all opcodes
to/from bytes between compile-time and runtime, let's just embed
necessary information within the opcode enum variants themselves. I'm
not sure why I didn't do this to begin with. This reduces the complexity
a ton, and probably also has runtime performance implications.

+++ abra_core/src/vm/prelude.rs
@@ -14,7 +14,7 @@ struct PreludeBinding {
 
 thread_local! {
   pub static PRELUDE: Prelude = Prelude::new();
-  pub static PRELUDE_NUM_CONSTS: u8 = PRELUDE.with(|p| p.num_bindings()) as u8;
+  pub static PRELUDE_NUM_CONSTS: usize = PRELUDE.with(|p| p.num_bindings());
   pub static PRELUDE_BINDINGS: Vec<(String, Value)> = PRELUDE.with(|p| p.get_bindings());
   pub static PRELUDE_BINDING_VALUES: Vec<Value> = PRELUDE.with(|p| p.get_binding_values());
 }
